+++ __init__.py
@@ -43,6 +43,12 @@ try:
     from .nodes.node_prompt_builder_backend import (
         NODE_DISPLAY_NAME_MAPPINGS as _NAMES_PROMPT_BUILDER,
     )
+    from .nodes.node_string_concatenate_backend import (
+        NODE_CLASS_MAPPINGS as _MAPS_STRING_CONCATENATE,
+    )
+    from .nodes.node_string_concatenate_backend import (
+        NODE_DISPLAY_NAME_MAPPINGS as _NAMES_STRING_CONCATENATE,
+    )
     from .nodes.node_save_image_avatary_backend import (
         NODE_CLASS_MAPPINGS as _MAPS_SAVE_IMAGE_ULTRA,
     )
@@ -79,6 +85,12 @@ except ImportError:
     )
     from nodes.node_prompt_builder_backend import (
         NODE_DISPLAY_NAME_MAPPINGS as _NAMES_PROMPT_BUILDER,
+    )
+    from nodes.node_string_concatenate_backend import (
+        NODE_CLASS_MAPPINGS as _MAPS_STRING_CONCATENATE,
+    )
+    from nodes.node_string_concatenate_backend import (
+        NODE_DISPLAY_NAME_MAPPINGS as _NAMES_STRING_CONCATENATE,
     )
     from nodes.node_save_image_avatary_backend import (
         NODE_CLASS_MAPPINGS as _MAPS_SAVE_IMAGE_ULTRA,
@@ -121,6 +133,7 @@ NODE_CLASS_MAPPINGS = _merge_mapping_dicts(
     _MAPS_NANO_BANANA,
     _MAPS_PROMPT_LIST,
     _MAPS_PROMPT_BUILDER,
+    _MAPS_STRING_CONCATENATE,
     _MAPS_SAVE_IMAGE_ULTRA,
     _MAPS_LOAD_IMAGE_BATCH,
     _MAPS_LORA_STACK,
@@ -135,6 +148,7 @@ NODE_DISPLAY_NAME_MAPPINGS = _merge_mapping_dicts(
     _NAMES_NANO_BANANA,
     _NAMES_PROMPT_LIST,
     _NAMES_PROMPT_BUILDER,
+    _NAMES_STRING_CONCATENATE,
     _NAMES_SAVE_IMAGE_ULTRA,
     _NAMES_LOAD_IMAGE_BATCH,
     _NAMES_LORA_STACK,