+++ web-src/web-build.ts
@@ -6,3 +6,4 @@ import "./features_avatary.js";
 import "./load_images_avatary.js";
 import "./lora_stack_frontend.js";
 import "./prompt_builder_frontend.js";
+import "./string_concatenate_frontend.js";